restruct the root commands file as it was starting to get unruly
