+++ cmd/version.go
@@ -1,0 +1,18 @@
+package cmd
+
+import (
+	"dp-cli/out"
+
+	"github.com/spf13/cobra"
+)
+
+func versionSubCommand() *cobra.Command {
+	return &cobra.Command{
+		Use:   "version",
+		Short: "Print the app version",
+		Run: func(cmd *cobra.Command, args []string) {
+			out.Info(appVersion)
+		},
+	}
+}
+
